fix: S12.29 compute Windows atomic counter wrap in uint32_t
The sequence-id increment in SolidSyslogWindowsAtomicCounter.c was performed
on a signed LONG (current + 1, plus a (uint32_t)-cast wrap guard), avoiding
signed overflow only because the guard fires exactly at INT32_MAX. Compute and
compare entirely in uint32_t, casting to LONG only at the
InterlockedCompareExchange boundary, matching the Std sibling so correctness no
longer depends on guard ordering. Behaviour is unchanged.

Adds the missing IncrementJustBelowMaxReturnsMax boundary case next to the
existing at-MAX wrap test as a regression guard.

Closes #530

+++ Platform/Windows/Source/SolidSyslogWindowsAtomicCounter.c
@@ -48,17 +48,17 @@ void WindowsAtomicCounter_Cleanup(struct SolidSyslogAtomicCounter* base)
 static uint32_t WindowsAtomicCounter_Increment(struct SolidSyslogAtomicCounter* base)
 {
     struct SolidSyslogWindowsAtomicCounter* self = WindowsAtomicCounter_SelfFromBase(base);
-    LONG current = InterlockedCompareExchange(&self->Value, 0, 0);
-    LONG next = 0;
+    uint32_t current = (uint32_t) InterlockedCompareExchange(&self->Value, 0, 0);
+    uint32_t next = 0U;
     do
     {
-        next = ((uint32_t) current >= SOLIDSYSLOG_SEQUENCE_ID_MAX) ? 1 : (current + 1);
-        LONG previous = InterlockedCompareExchange(&self->Value, next, current);
+        next = (current >= SOLIDSYSLOG_SEQUENCE_ID_MAX) ? 1U : (current + 1U);
+        uint32_t previous = (uint32_t) InterlockedCompareExchange(&self->Value, (LONG) next, (LONG) current);
         if (previous == current)
         {
             break;
         }
         current = previous;
     } while (1);
-    return (uint32_t) next;
+    return next;
 }
